[core] project multiple types to int

+++ saltify-lagrange-core/src/main/kotlin/org/ntqqrev/saltify/lagrange/packet/message/element/CommonElement.kt
@@ -6,5 +6,5 @@ import org.ntqqrev.saltify.protobuf.annotation.ProtoField
 class CommonElement(
     @ProtoField(1) var serviceType: Int,
     @ProtoField(2) var pbElem: ByteArray,
-    @ProtoField(3) var businessType: Long,
+    @ProtoField(3) var businessType: Int,
 ) : ProtoMessage()
